refactor: update categoryMapping import paths
- Changed import paths for `categoryMapping` from the lib directory to the new utils directory in multiple components, enhancing code organization and maintainability.
- Introduced a new `category-mapping.ts` file to centralize category mappings.
