+++ src/components/conveyor-item.tsx
@@ -1,11 +1,11 @@
 import Image from "next/image";
 import { createFilterSchema } from "@/schemas/filterFormSchema";
+import { categoryMapping } from "@/utils/category-mapping";
 import { XIcon } from "lucide-react";
 import { Control, useFormContext } from "react-hook-form";
 import { z } from "zod";
 
 import { type ItemWithFields } from "@/types/item";
-import { categoryMapping } from "@/lib/categoryMapping";
 import { FilterSettingsFieldDescription } from "@/lib/constants";
 import { Button } from "@/components/ui/button";
 import { getCategoryIcon } from "@/components/category-icons";
