+++ src/components/filters/filter-items-carousel.tsx
@@ -2,12 +2,12 @@
 
 import * as React from "react";
 import Image from "next/image";
+import { categoryMapping } from "@/utils/category-mapping";
 import Autoplay from "embla-carousel-autoplay";
 import { useInView } from "react-intersection-observer";
 
 import type { ConveyorFilterItem } from "@/types/filter";
 import { useMediaQuery } from "@/hooks/use-media-query";
-import { categoryMapping } from "@/lib/categoryMapping";
 import { Card, CardContent } from "@/components/ui/card";
 import {
   Carousel,
