+++ src/components/filters/view-filter.tsx
@@ -1,11 +1,11 @@
 import { useState } from "react";
 import Image from "next/image";
+import { categoryMapping } from "@/utils/category-mapping";
 import { EyeIcon } from "lucide-react";
 
 import type { ConveyorFilter, ConveyorFilterItem } from "@/types/filter";
 import { useLogFilterEvent } from "@/hooks/use-log-filter-event";
 import { useMediaQuery } from "@/hooks/use-media-query";
-import { categoryMapping } from "@/lib/categoryMapping";
 import { Button } from "@/components/ui/button";
 import { ButtonWithIcon } from "@/components/ui/button-with-icon";
 import { Card } from "@/components/ui/card";
