+++ src/utils/category-mapping.ts
